+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Fixed
 
 - [TD-5732] Fixed reindex all implementation versions when add/delete links
+- [TD-5442] Fix hard delete implementation versions
 
 ## [5.5.1] 2023-04-18